2.0
- Added conditions to check if Gravity Forms is activated.
- Added admin notification, when this add-on is active, but Gravity Forms is not active.

+++ surbma-gdpr-proof-gravity-forms.php
@@ -5,7 +5,7 @@ Plugin Name: Surbma - GDPR Proof Gravity Forms
 Plugin URI: https://surbma.com/wordpress-plugins/
 Description: Makes your Gravity Forms forms GDPR compatible.
 
-Version: 1.0
+Version: 2.0
 
 Author: Surbma
 Author URI: https://surbma.com/
@@ -20,9 +20,18 @@ Domain Path: /languages/
 if ( !defined( 'ABSPATH' ) ) exit( 'Good try! :)' );
 
 // Prevent GF to save visitor IP address
-add_filter( 'gform_ip_address', '__return_empty_string' );
+if ( class_exists( 'GFForms' ) )
+	add_filter( 'gform_ip_address', '__return_empty_string' );
 
 function surbma_gpgf_remove_form_entry( $entry ) {
-	GFAPI::delete_entry( $entry['id'] );
+	if ( class_exists( 'GFForms' ) )
+		GFAPI::delete_entry( $entry['id'] );
 }
 add_action( 'gform_after_submission', 'surbma_gpgf_remove_form_entry' );
+
+function surbma_gpgf_wp_admin_notices() {
+	if ( !class_exists( 'GFForms' ) ) {
+		_e( '<div class="error notice"><p><strong>Gravity Forms is not activated!</strong> This plugin is an add-on for Gravity Forms plugin. You need to activate Gravity Forms plugin before using this plugin.</p></div>', 'surbma-gdpr-proof-gravity-forms' );
+	}
+}
+add_action( 'admin_notices', 'surbma_gpgf_wp_admin_notices' );
